# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmniReferral API",
        "description": "REST API for the OmniReferral real estate referral and lead management platform.\n\nProvides authentication, user management, realtor profiles, GoHighLevel CRM integration,\nonboarding webhooks, and administrative functions.\n\n## Authentication\nMost endpoints require a Bearer token obtained via the Login or Register endpoints.\nPass the token in the `Authorization` header as `Bearer <token>`.\n\n## Webhooks\nGoHighLevel webhook endpoints are public (CSRF exempt) and authenticated via\nthe `X-OmniReferral-Webhook` header secret.",
        "contact": {
            "email": "admin@omnireferrals.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://omnireferrals.com",
            "description": "OmniReferral Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/admin/email": {
            "get": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "View email and auth diagnostics",
                "description": "Shows current mail configuration, delivery stats, email logs, and auth activity. Admin only.",
                "operationId": "22283c6e9b5565754340091e053b00c6",
                "responses": {
                    "200": {
                        "description": "Email diagnostics page"
                    },
                    "403": {
                        "description": "Forbidden - admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/email/test": {
            "post": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "Send test email from admin panel",
                "description": "Sends a plain text test email to verify mail delivery. Admin only.",
                "operationId": "add56453d6e39cfdce0b3c0148c3eaba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/email/smtp-test": {
            "post": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "Test SMTP connection from admin panel",
                "description": "Tests the TCP connection to the configured SMTP server. Admin only.",
                "operationId": "91bd6632d3d20bc4b06e1efbab8fe861",
                "responses": {
                    "200": {
                        "description": "Connection test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "sanctum",
                "description": "Enter Sanctum token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}